Error control

+++ app.py
@@ -136,7 +136,10 @@ def register():
             msg = Message('Confirm_Email', recipients=[name])
             link = url_for('confirm_email', token=token, external=True)
             msg.body = 'Your link is https://haveyouseenit.herokuapp.com{}'.format(link)
-            mail.send(msg)
+            try : 
+                mail.send(msg)
+            except :
+                return apology("provide a valid Email-ID")
             db.execute("INSERT INTO users(username, hash, tablename, confirmation) VALUES(?, ?, ?, 0);", name, generate_password_hash(password), name)
             db.execute("INSERT INTO personalinfo(id, profilepic) VALUES(?, 0);", db.execute("SELECT id FROM users WHERE username = ?", name)[0]['id'])
             db.execute("CREATE TABLE ?(id INTEGER NOT NULL , name TEXT, seen INTEGER, total INTEGER, type TEXT NOT NULL, time DATETIME, time_per_episode FLOAT, times INTEGER, PRIMARY KEY(id))",name)
@@ -152,9 +155,9 @@ def register():
 
 @app.route('/confirm_email/<token>')
 def confirm_email(token):
-    email = s.loads(token, salt = 'email-confirm', max_age = 3600)
+    email = s.loads(token, salt = 'email-confirm', max_age = 360000)
     db.execute("UPDATE users SET confirmation = 1 WHERE username =?", email)
-    return 'Your Email Has been confirmed'
+    return render_template("confirmemail.html")
 
 @app.route('/addupdate', methods = ['GET', 'POST'])
 @login_required
@@ -205,7 +208,12 @@ def search():
         return render_template("search.html")
     if request.method == 'POST':
         name = request.form.get("name")
-        return render_template("results.html", results=lookup(name)) 
+        if not name:
+            return apology("Type Something")
+        results=lookup(name)
+        if not results:
+            return apology("No results")
+        return render_template("results.html", results=results) 
 
 @app.route("/searchanime", methods=["GET", "POST"])
 @login_required
@@ -214,7 +222,12 @@ def searchanime():
         return render_template("search.html")
     if request.method == 'POST':
         name = request.form.get("name")
-        return render_template("resultsanime.html", results=lookupanime(name))
+        if not name:
+            return apology("Type Something")
+        results=lookupanime(name)
+        if not results:
+            return apology("No results")
+        return render_template("resultsanime.html", results=results)
 
 @app.route("/changepass", methods=["GET", "POST"])
 @login_required
@@ -239,6 +252,8 @@ def change():
     if request.method == "POST":
         column = request.form.get("column")
         data = request.form.get("data")
+        if column == "Choose..." :
+            return apology("Select Column")
         db.execute("UPDATE personalinfo SET ? = ? WHERE id = ?", column, data, session['user_id'])
         return render_template("change.html", personalinfo=db.execute("SELECT * FROM personalinfo WHERE id = ?", session['user_id'])[0])
 
@@ -256,16 +271,20 @@ def myself():
 @login_required
 def update():
     id = request.form.get('id')
-    seen = request.form.get('seen')
+    value = request.form.get('seen')
     type = request.form.get('type')
+    if not id or value == "Choose..." or type == "Choose...":
+        return apology("Select Columns And Enter Values")
     name = db.execute("SELECT username FROM users WHERE id = ?", session['user_id'])[0]['username']
-    db.execute("UPDATE ? SET ? = ? WHERE id = ?", name, type, seen, id)
+    db.execute("UPDATE ? SET ? = ? WHERE id = ?", name, type, value, id)
     return render_template("addupdate.html", table=db.execute("SELECT * FROM ?", name))
 
 @app.route('/delete', methods = ["POST"])
 @login_required
 def delete():
     id = request.form.get('id')
+    if id == "Choose...":
+        return apology("Select ID")
     name = db.execute("SELECT username FROM users WHERE id = ?", session['user_id'])[0]['username']
     db.execute("DELETE FROM ? WHERE id = ?", name, id)
     return render_template("addupdate.html", table=db.execute("SELECT * FROM ?", name))
@@ -284,7 +303,12 @@ def usersearch():
         return render_template("usrsrch.html")
     if request.method == "POST":
         name = request.form.get("username")
+        if not name:
+            return apology("Enter Username")
+        try:
+            id = db.execute("SELECT id FROM users WHERE username = ?", name)[0]['id']
+        except:
+            return apology("Enter Correct Email ID")
         timee=db.execute("SELECT sum((time_per_episode * seen)) AS typesum from ? group by type ORDER BY typesum DESC LIMIT 1;", name)[0]['typesum']
         length = float(2.5 ** (len(str(timee)) - 2))
-        id = db.execute("SELECT id FROM users WHERE username = ?", name)[0]['id']
         return render_template("userresult.html", table=db.execute("SELECT * FROM ?", name), personalinfo=db.execute("SELECT * FROM personalinfo WHERE id = ?", id)[0], time_type=db.execute("SELECT sum((time_per_episode * seen)*times) / ? AS typesum, sum((time_per_episode * seen)*times)/60.0 as types, type, type from ? group by type;", length, name), total_time=db.execute("SELECT sum((time_per_episode * seen * times)/60) AS TOTALSUM from ?;", name)[0])
